pyproject.toml
[project]
name = "join-the-siege"
version = "0.1.0"
description = ""
authors = [
    {name = "Nakul",email = "nakulj2@illinois.edu"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "joblib (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
